requirements.txt
face_recognition==1.3.0
dlib==19.22.0
opencv-python==4.11.0.86
elasticsearch==8.13.2
psycopg2-binary==2.9.10
numpy==2.2.6

